Fix memory issues with transactions
Transactions cause the local db to be changed, this invalidates Packages
that point to the localdb. This means commiting a transaction needs
a mutable reference to ensure there are no other package references in
scope.

However, when building the transaction, we need a shared reference so
packages can be inserted into it.

To do this, transaction functions have been moved to the handle. This
makes things slightly less erganomic and the trans now needs to be freed
by the user after use.